content/svg/content/test/test_pointer-events-3.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=762679
-->
<head>
  <title>Test pointer events for small objects scaled up</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="run()">
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function run()
{
  var svgDoc = document.getElementById("svg");
  var div = document.getElementById("div");
  // Get the coords of the origin of the SVG canvas:
  var originX = div.offsetLeft;
  var originY = div.offsetTop;
  var circle = document.getElementById("circle");

  var elementFromPoint = document.elementFromPoint(originX + 150, originY + 51);
  is(elementFromPoint, circle, "Top of circle should hit");

  var elementFromPoint = document.elementFromPoint(originX + 249, originY + 150);
  is(elementFromPoint, circle, "Right of circle should hit");

  var elementFromPoint = document.elementFromPoint(originX + 150, originY + 249);
  is(elementFromPoint, circle, "Bottom of circle should hit");

  var elementFromPoint = document.elementFromPoint(originX + 51, originY + 150);
  is(elementFromPoint, circle, "Left of circle should hit");

  SimpleTest.finish();
}

]]>
</script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=762679">Mozilla Bug 762679</a>
<p id="display"></p>
<div id="content">

  <div width="100%" height="1" id="div">
  </div>
  <svg xmlns="http://www.w3.org/2000/svg" width="100%" height="400" id="svg">
    <circle id="circle" cx="1.5" cy="1.5" r="1" transform="scale(100, 100)"/>
  </svg>

</div>
<pre id="test">
</pre>
</body>
</html>
